cambiada la estructura a typescript e instalado dependencia para mongoose
