+++ package.json
@@ -9,7 +9,8 @@
     },
     "dependencies": {
         "dotenv": "^16.4.5",
-        "express": "^4.19.2"
+        "express": "^4.19.2",
+        "mongoose": "^8.8.1"
     },
     "devDependencies": {
         "@types/express": "^4.17.21",
