+++ src/controllers/userController.ts
@@ -1,0 +1,25 @@
+import { Request, Response } from 'express';
+import Usuario from '../models/userModel';
+
+export const loginUser = async (req: Request, res: Response): Promise<void> => {
+    const { email, password } = req.body;
+
+    try {
+        const user = await Usuario.findOne({ email });
+        if (!user) {
+            res.status(404).json({ message: 'Usuario no encontrado' });
+            return;
+        }
+
+        if (user.password !== password) {
+            res.status(400).json({ message: 'Contraseña incorrecta' });
+            return;
+        }
+
+        res.status(200).json({ message: 'Inicio de sesión exitoso', user });
+    } catch (error) {
+        res.status(500).json({ message: 'Error al iniciar sesión', error });
+    }
+};
+
+// Pendiente agregar seguridad al password
