--- src/routes/userRoutes.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const { loginUser } = require('../controllers/userController');
-const router = express.Router();
-
-router.post('/login', loginUser); // Ruta de login
-
-module.exports = router;
